Fix FK type mismatch: user_profiles.user_id INT(11) to match users.id

+++ src/main/resources/db/migration/V6_1__create_user_profiles_table.sql
@@ -1,7 +1,7 @@
 -- Table des profils employés (un compte entreprise → plusieurs profils)
 CREATE TABLE user_profiles (
     id           BIGINT       AUTO_INCREMENT PRIMARY KEY,
-    user_id      BIGINT       NOT NULL,
+    user_id      INT(11)      NOT NULL,
     display_name VARCHAR(100) NOT NULL,
     color        VARCHAR(20)  NOT NULL DEFAULT '#3b82f6',
     active       BOOLEAN      NOT NULL DEFAULT TRUE,
